+++ src/com/horovod/timecountfxprobe/user/Manager.java
@@ -11,7 +11,7 @@ public class Manager implements User {
     private Role role = Role.MANAGER;
     private String fullName;
     private String email;
-    private double workHourValue = 0.0;
+    private int workHourValue = 0;
     private boolean isRetired = false;
 
     public Manager(String nameLogin, String password) {
@@ -66,11 +66,11 @@ public class Manager implements User {
     }
 
     @XmlElement(name = "managerworkhourvalue")
-    public double getWorkHourValue() {
+    public int getWorkHourValue() {
         return workHourValue;
     }
 
-    public void setWorkHourValue(double workHourValue) {
+    public void setWorkHourValue(int workHourValue) {
         this.workHourValue = workHourValue;
     }
