+++ src/com/horovod/timecountfxprobe/user/User.java
@@ -11,8 +11,8 @@ public interface User {
     void setFullName(String newFullName);
     String getEmail();
     void setEmail(String newEmail);
-    double getWorkHourValue();
-    void setWorkHourValue(double workHourValue);
+    int getWorkHourValue();
+    void setWorkHourValue(int workHourValue);
     boolean isRetired();
     void setRetired(boolean retired);
 }
